Use a resolver object rather than a method
