+++ examples/universal-redux/src/genStore.js
@@ -1,7 +1,4 @@
-import {
-  createHistoryEnhancer,
-  queryMiddleware,
-} from 'farce';
+import { createHistoryEnhancer, queryMiddleware } from 'farce';
 import createMatchEnhancer from 'found/lib/createMatchEnhancer';
 import foundReducer from 'found/lib/foundReducer';
 import Matcher from 'found/lib/Matcher';
@@ -10,7 +7,7 @@ import { combineReducers, compose, createStore } from 'redux';
 import routeConfig from './routeConfig';
 
 export default function genStore(historyProtocol, preloadedState) {
-  const store = createStore(
+  return createStore(
     combineReducers({
       found: foundReducer,
     }),
@@ -25,6 +22,4 @@ export default function genStore(historyProtocol, preloadedState) {
       ),
     ),
   );
-
-  return store;
 }
