+++ src/createInitialBrowserRouter.js
@@ -1,12 +1,12 @@
 import BrowserProtocol from 'farce/lib/BrowserProtocol';
 
 import createInitialFarceRouter from './createInitialFarceRouter';
-import resolveElements from './resolveElements';
+import resolver from './resolver';
 
 export default async function createInitialBrowserRouter(options) {
   return createInitialFarceRouter({
     ...options,
     historyProtocol: new BrowserProtocol(),
-    resolveElements,
+    resolver,
   });
 }
