+++ src/index.js
@@ -22,7 +22,7 @@ export Matcher from './Matcher';
 export { matcherShape, matchShape, routerShape } from './PropTypes';
 export Redirect from './Redirect';
 export RedirectException from './RedirectException';
-export resolveElements from './resolveElements';
+export resolver from './resolver';
 export ResolverUtils from './ResolverUtils';
 export Route from './Route';
 export withRouter from './withRouter';
